backend/pyproject.toml
[tool.poetry]
name = "email-assistant-backend"
version = "0.1.0"
description = "AI-Based Email Assistant"
authors = ["Frank Bria <frank@frankbria.com>"]
license = "mit"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
python-dotenv = "^1.1.0"
pydantic = "^2.11.3"
beanie = "^1.29.0"
pydantic-settings = "^2.9.1"
openai = "^1.76.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
ruff = "^0.11.6"
mypy = "^1.15.0"
pytest-cov = "^6.1.1"
pytest-asyncio = "^0.26.0"
pytest-timeout = "^2.3.1"

[tool.pytest.ini_options]
pythonpath = [
    ".",
    "app",
    "../backend"
]
testpaths = ["tests"]
addopts = "-v"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
